DSEC: optical flow pngs need to be floats

+++ tonic/datasets/dsec.py
@@ -288,7 +288,7 @@ class DSEC(Dataset):
                 png_filenames = list_files(full_base_folder, ".png", prefix=True)
                 target = np.array(
                     [imageio.imread(file, format="PNG-FI") for file in png_filenames]
-                )
+                ).astype(float)
                 target[:, :, :, :2] -= 2 ^ 15
                 target[:, :, :, :2] /= 128
